# docs/swagger.yaml
basePath: /api/v1
definitions:
  ErrorResponse:
    properties:
      error:
        example: invalid_request
        type: string
      message:
        example: The request was invalid
        type: string
      status:
        example: 400
        type: integer
    type: object
...
  SuccessResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      status:
        example: 200
        type: integer
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@themonkeys.com
    name: The Monkeys Team
    url: https://github.com/the-monkeys
  description: A comprehensive IAM system providing authentication, authorization,
    and access control for multi-tenant organizations with policy-based security,
    session management, and comprehensive audit trails.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://themonkeys.com/terms
  title: Monkeys Identity & Access Management API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of users with filtering and sorting options
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Sort field (default: created_at)'
        in: query
        name: sort
        type: string
      - description: 'Sort order: asc or desc (default: desc)'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved users list
          schema:
            $ref: '#/definitions/SuccessResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - User Management
    post:
      consumes:
      - application/json
      description: Create a new user account with the provided details
      parameters:
      - description: User creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SuccessResponse'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - User Management
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user account by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - User Management
    get:
      consumes:
      - application/json
      description: Retrieve a specific user's details by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - User Management
    put:
      consumes:
      - application/json
      description: Update a user's profile information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SuccessResponse'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid request format or user ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - User Management
  /users/{id}/activate:
    post:
      consumes:
      - application/json
      description: Activate a previously suspended user account
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User activated successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Activate user
      tags:
      - User Management
  /users/{id}/profile:
    get:
      consumes:
      - application/json
      description: Retrieve detailed profile information for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user profile
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: User profile not found
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User Management
    put:
      consumes:
      - application/json
      description: Update specific profile fields for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Profile update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SuccessResponse'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid request format or user ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User Management
  /users/{id}/sessions:
    delete:
      consumes:
      - application/json
      description: Revoke all active sessions for a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User sessions revoked successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke user sessions
      tags:
      - User Management
    get:
      consumes:
      - application/json
      description: Retrieve all active sessions for a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user sessions
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user sessions
      tags:
      - User Management
  /users/{id}/suspend:
    post:
      consumes:
      - application/json
      description: Suspend a user account with a reason
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Suspension details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SuccessResponse'
      produces:
      - application/json
      responses:
        "200":
          description: User suspended successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid request format or user ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Suspend user
      tags:
      - User Management
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"